conv_lstm_satellite/requirements.txt
torch==1.13.0
numpy==1.21.6
Pillow==9.2.0
opencv-python==4.6.0.66
scikit-learn==1.1.3
matplotlib==3.5.3
yaml==6.0
tqdm==4.64.0